+++ src/backend/tfjs_backend_test.ts
@@ -1179,35 +1179,6 @@ describeMathCPUAndGPU('dropout', () => {
   }
 });
 
-describeMathCPUAndGPU('l2Normalize', () => {
-  it('normalizes with no axis defined.', () => {
-    const x = tensor2d([[1, 2], [3, 4]], [2, 2]);
-    const norm = Math.sqrt(1 * 1 + 2 * 2 + 3 * 3 + 4 * 4);
-    const expected =
-      tensor2d([[1 / norm, 2 / norm], [3 / norm, 4 / norm]], [2, 2]);
-    const result = K.l2Normalize(x);
-    expectTensorsClose(result, expected);
-  });
-
-  it('normalizes along axis = -1.', () => {
-    const x = tensor2d([[1, 2], [3, 4]], [2, 2]);
-    const firstNorm = Math.sqrt(1 * 1 + 2 * 2);
-    const secondNorm = Math.sqrt(3 * 3 + 4 * 4);
-    const expected = tensor2d(
-      [[1 / firstNorm, 2 / firstNorm], [3 / secondNorm, 4 / secondNorm]],
-      [2, 2]);
-    const result = K.l2Normalize(x, -1);
-    expectTensorsClose(result, expected);
-  });
-
-  it('normalizes with zeros.', () => {
-    const x = zeros([2, 2]);
-    const result = K.l2Normalize(x);
-    expectTensorsClose(result, x);
-  });
-});
-
-
 describeMathCPUAndGPU('biasAdd', () => {
   it('1D + 1D', () => {
     const x = tfc.ones([2]);
